modules: move LmodFileLayout defs into BaseFileLayout
Move hierarchical-specific methods of LmodFileLayout class into generic
BaseFileLayout to allow hierarchical setup also for Tcl.

Adapt "arch_dirname" method to better rely on it in "filename" method.
