+++ lib/spack/spack/modules/common.py
@@ -28,6 +28,8 @@ import collections
 import contextlib
 import copy
 import datetime
+import inspect
+import itertools
 import os
 import re
 import string
@@ -40,7 +42,7 @@ import spack.config
 import spack.deptypes as dt
 import spack.environment
 import spack.error
-import spack.llnl.util.filesystem
+import spack.llnl.util.filesystem as fs
 import spack.llnl.util.tty as tty
 import spack.paths
 import spack.projections as proj
@@ -226,7 +228,7 @@ def generate_module_index(root, modules, overwrite=False):
         entry = {"path": m.layout.filename, "use_name": m.layout.use_name}
         entries[m.spec.dag_hash()] = entry
     index = {"module_index": entries}
-    spack.llnl.util.filesystem.mkdirp(root)
+    fs.mkdirp(root)
     with open(index_path, "w", encoding="utf-8") as index_file:
         syaml.dump(index, default_flow_style=False, stream=index_file)
 
@@ -548,20 +550,157 @@ class BaseFileLayout:
         return "-".join(path_elements)
 
     @property
+    def arch_dirname(self):
+        """Returns the root folder for THIS architecture"""
+        # Architecture sub-folder
+        arch_folder_conf = spack.config.get("modules:%s:arch_folder" % self.conf.name, True)
+        if arch_folder_conf:
+            # include an arch specific folder between root and filename
+            if self.conf.hierarchical:
+                arch_folder = "-".join(
+                    [str(self.spec.platform), str(self.spec.os), str(self.spec.target.family)]
+                )
+            else:
+                arch_folder = str(self.spec.architecture)
+            return os.path.join(self.dirname(), arch_folder)
+        return self.dirname()
+
+    @property
     def filename(self):
         """Name of the module file for the current spec."""
         # Just the name of the file
         filename = self.use_name
         if self.extension:
             filename = f"{self.use_name}.{self.extension}"
-        # Architecture sub-folder
-        arch_folder_conf = spack.config.get("modules:%s:arch_folder" % self.conf.name, True)
-        if arch_folder_conf:
-            # include an arch specific folder between root and filename
-            arch_folder = str(self.spec.architecture)
-            filename = os.path.join(arch_folder, filename)
+
+        if self.conf.hierarchical:
+            # Get the list of requirements and build an **ordered**
+            # list of the path parts
+            requires = self.conf.requires
+            hierarchy = self.conf.hierarchy_tokens
+            path_parts = lambda x: self.token_to_path(x, requires[x])
+            parts = [path_parts(x) for x in hierarchy if x in requires]
+
+            # My relative path if just a join of all the parts
+            hierarchy_name = os.path.join(*parts)
+
+            filename = os.path.join(hierarchy_name, filename)
+
         # Return the absolute path
-        return os.path.join(self.dirname(), filename)
+        return os.path.join(self.arch_dirname, filename)
+
+    def token_to_path(self, name, value):
+        """Transforms a hierarchy token into the corresponding path part.
+
+        Args:
+            name (str): name of the service in the hierarchy
+            value: actual provider of the service
+
+        Returns:
+            str: part of the path associated with the service
+        """
+
+        # General format for the path part
+        def path_part_fmt(token):
+            return fs.polite_path([f"{token.name}", f"{token.version}"])
+
+        # If we are dealing with a core compiler, return 'Core'
+        core_compilers = self.conf.core_compilers
+        if name == "compiler" and any(spack.spec.Spec(value).satisfies(c) for c in core_compilers):
+            return "Core"
+
+        # Spec does not have a hash, as we are not allowed to
+        # use different flavors of the same compiler
+        if name == "compiler":
+            return path_part_fmt(token=value)
+
+        # In case the hierarchy token refers to a virtual provider
+        # we need to append a hash to the version to distinguish
+        # among flavors of the same library (e.g. openblas~openmp vs.
+        # openblas+openmp)
+        return f"{path_part_fmt(token=value)}-{value.dag_hash(length=7)}"
+
+    @property
+    def available_path_parts(self):
+        """List of path parts that are currently available. Needed to
+        construct the file name.
+        """
+        # List of available services
+        available = self.conf.available
+        # List of services that are part of the hierarchy
+        hierarchy = self.conf.hierarchy_tokens
+        # Tokenize each part that is both in the hierarchy and available
+        return [self.token_to_path(x, available[x]) for x in hierarchy if x in available]
+
+    @property
+    @memoized
+    def unlocked_paths(self):
+        """Returns a dictionary mapping conditions to a list of unlocked
+        paths.
+
+        The paths that are unconditionally unlocked are under the
+        key 'None'. The other keys represent the list of services you need
+        loaded to unlock the corresponding paths.
+        """
+
+        unlocked = collections.defaultdict(list)
+
+        # Get the list of services we require and we provide
+        requires_key = list(self.conf.requires)
+        provides_key = list(self.conf.provides)
+
+        # A compiler is always required. To avoid duplication pop the
+        # 'compiler' item from required if we also **provide** one
+        if "compiler" in provides_key:
+            requires_key.remove("compiler")
+
+        # Compute the unique combinations of the services we provide
+        combinations = []
+        for ii in range(len(provides_key)):
+            combinations += itertools.combinations(provides_key, ii + 1)
+
+        # Attach the services required to each combination
+        to_be_processed = [x + tuple(requires_key) for x in combinations]
+
+        # Compute the paths that are unconditionally added
+        # and append them to the dictionary (key = None)
+        available_combination = []
+        for item in to_be_processed:
+            hierarchy = self.conf.hierarchy_tokens
+            available = self.conf.available
+            ac = [x for x in hierarchy if x in item]
+            available_combination.append(tuple(ac))
+            parts = [self.token_to_path(x, available[x]) for x in ac]
+            unlocked[None].append(tuple([self.arch_dirname] + parts))
+
+        # Deduplicate the list
+        unlocked[None] = list(dedupe(unlocked[None]))
+
+        # Compute the combination of missing requirements: this will lead to
+        # paths that are unlocked conditionally
+        missing = self.conf.missing
+
+        missing_combinations = []
+        for ii in range(len(missing)):
+            missing_combinations += itertools.combinations(missing, ii + 1)
+
+        # Attach the services required to each combination
+        for m in missing_combinations:
+            to_be_processed = [m + x for x in available_combination]
+            for item in to_be_processed:
+                hierarchy = self.conf.hierarchy_tokens
+                available = self.conf.available
+                token2path = lambda x: self.token_to_path(x, available[x])
+                parts = []
+                for x in hierarchy:
+                    if x not in item:
+                        continue
+                    value = token2path(x) if x in available else x
+                    parts.append(value)
+                unlocked[m].append(tuple([self.arch_dirname] + parts))
+            # Deduplicate the list
+            unlocked[m] = list(dedupe(unlocked[m]))
+        return unlocked
 
 
 class BaseContext(tengine.Context):
@@ -927,7 +1066,7 @@ class BaseModuleFileWriter:
         # create it
         module_dir = os.path.dirname(self.layout.filename)
         if not os.path.exists(module_dir):
-            spack.llnl.util.filesystem.mkdirp(module_dir)
+            fs.mkdirp(module_dir)
 
         # Get the template for the module
         template_name = self._get_template()
